Working on dvb subtitles
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@2354 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Logic/TransportStream/DvbSubPes.cs
@@ -351,12 +351,13 @@ namespace Nikse.SubtitleEdit.Logic.TransportStream
             }
         }
 
-        public int PresentationTimeStampToMilliseconds()
+        public ulong PresentationTimeStampToMilliseconds()
         {
             if (PresentationTimeStamp.HasValue)
-                return (int)Math.Round(PresentationTimeStamp.Value + 45.0 / 90.0);
+                return (ulong)Math.Round(PresentationTimeStamp.Value + 45.0 / 90.0);
             else
                 return 0;
         }
+     
     }
 }
